pyproject.toml
[project]
name = "practicalapproachtoagenticai"
version = "0.1.0"
description = "The codebase for the book 'Practical Approach to Agentic AI'"
authors = [
    {name = "Vikas Gautam",email = "vikas.gautam18@gmail.com"},
    {name = "Shweta Kamath",email = "shwetkmth@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
License = "MIT"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "Gradio (>=5.31.0,<6.0.0)",
    "ollama (>=0.4.8,<0.5.0)",
    "semantic-kernel (==1.33.0)",
    "requests (>=2.32.4,<3.0.0)",
    "crewai (>=0.141.0,<0.142.0)",
    "langgraph (>=0.5.3,<0.6.0)",
    "langchain[openai] (>=0.3.26,<0.4.0)",
    "langgraph-swarm (>=0.0.13,<0.0.14)",
    "autogen (>=0.9.7,<0.10.0)",
    "autogen-agentchat (>=0.7.1,<0.8.0)",
    "autogen-ext[azure] (>=0.7.1,<0.8.0)",
    "azure-ai-inference (>=1.0.0b9,<2.0.0)",
    "autogen-core (>=0.7.1,<0.8.0)",
    "mcp[cli] (>=1.13.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "chainlit (>=2.7.1.1,<3.0.0.0)",
    "agent-framework (>=1.0.0b251016,<2.0.0)",
    "agent-framework-core (>=1.0.0b251016,<2.0.0)",
    "openai (==1.109.1)",
    "openai-agents (==0.3.3)",
    "shutup (>=0.2.0,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
